+++ helpers/handlebars/helpers/index.js
@@ -3,15 +3,10 @@ const moment = require('moment');
 const truncatehtml = require('truncate-html');
 const stripHtml = require('string-strip-html');
 const { Configuration } = require('@hpi-schul-cloud/commons');
-const path = require('path');
-const fs = require('fs');
 const { getStaticAssetPath } = require('../../../middleware/assets');
 const permissionsHelper = require('../../permissions');
 const i18n = require('../../i18n');
 const Globals = require('../../../config/global');
-
-const themeAssetDir = path.join(__dirname, '../../..', `build/${Globals.SC_THEME}`);
-
 
 const timesHelper = require('../../timesHelper');