+++ CHANGELOG.md
@@ -7,14 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Security`
 
-## Added
+## Unreleased
 
-- SC-8246 - Add possible to fingerprint sset if FEATURE_ASSET_CACHING_ENABLED is true
+### Added
+
+- SC-8246 - Add fingerprints to static assets when FEATURE_ASSET_CACHING_ENABLED is true for assets referenced within of hbs files
 
 ### Fixed
 
 - SC-5710 - Updated the accessibility of the course page with fixed delete button for course time (for keyboard users)
-
 
 ## 25.4.0